chore: LayoutController, PipeRecipesController
