+++ packages/autopilot/src/app/controllers/app-menu.ts
@@ -22,12 +22,12 @@ import { ViewportManager } from '../managers/viewport-manager';
 import { ProjectController } from './project';
 import { dom, helpers } from '../util';
 import { Viewport } from '../viewport';
-import { LayoutManager } from '../managers/layout-manager';
 import { PlaybackManager } from '../managers/playback-manager';
 import { EmulationController } from './emulation';
 import { ToolsController } from './tools';
 import { controller } from '../controller';
 import { ProtocolController } from './protocol';
+import { LayoutController } from './layout';
 
 const UI_HIDE_MENU = booleanConfig('UI_HIDE_MENU', true);
 
@@ -45,8 +45,8 @@ export class AppMenuController {
         protected settings: SettingsController,
         @inject(ViewportManager)
         protected viewports: ViewportManager,
-        @inject(LayoutManager)
-        protected layout: LayoutManager,
+        @inject(LayoutController)
+        protected layout: LayoutController,
         @inject(ProjectController)
         protected project: ProjectController,
         @inject(PlaybackManager)
