+++ packages/autopilot/src/app/controllers/index.ts
@@ -26,6 +26,8 @@ export * from './extension-loader';
 export * from './extension-registry';
 export * from './first-run';
 export * from './inspect';
+export * from './layout';
+export * from './pipe-recipes';
 export * from './project';
 export * from './protocol';
 export * from './roxi';
