+++ packages/autopilot/src/app/provider.ts
@@ -37,8 +37,10 @@ export function createControllerProvider(app: App) {
         extReg: app.get(ctl.ExtensionRegistryController),
         firstRun: app.get(ctl.FirstRunController),
         inspect: app.get(ctl.InspectController),
+        layout: app.get(ctl.LayoutController),
         project: app.get(ctl.ProjectController),
         protocol: app.get(ctl.ProtocolController),
+        pipeRecipes: app.get(ctl.PipeRecipesController),
         roxi: app.get(ctl.RoxiController),
         scriptReport: app.get(ctl.ScriptReportController),
         settings: app.get(ctl.SettingsController),
